+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fix `yt articles tree` command throwing NoneType error when parentArticle field is null (#299)
+
 ## [0.10.0] - 2025-07-20
 
 ### Added